+++ addons/rbu/$PBOPREFIX$
@@ -1,0 +1,1 @@
+y\jtog\addons\rbu
