+++ addons/rbu/functions/script_component.hpp
@@ -1,0 +1,1 @@
+#include "\y\jtog\addons\rbu\script_component.hpp"
